+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     `undefined`). This would issue a lazyLoad event and a resopnse of `[]` would
     mark the node as leaf node.
   * [Added] ext-filter methods node.isMatched() and tree.isFilterActive()
+  * [Added] CSS for ext-childcounter badges is now part of the themes
 
 # 2.12.0 / 2015-09-10
   * [Changed] Documented `iconClass` callback and changed signature from